[refactor] All OpenCL schedulers moved to scheduler package
