+++ tornado-drivers/opencl/src/main/java/uk/ac/manchester/tornado/drivers/opencl/scheduler/OCLAMDScheduler.java
@@ -2,7 +2,7 @@
  * This file is part of Tornado: A heterogeneous programming framework:
  * https://github.com/beehive-lab/tornadovm
  *
- * Copyright (c) 2013-2020, APT Group, Department of Computer Science,
+ * Copyright (c) 2020, APT Group, Department of Computer Science,
  * The University of Manchester. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
@@ -21,35 +21,46 @@
  * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
  *
  */
-package uk.ac.manchester.tornado.drivers.opencl;
+package uk.ac.manchester.tornado.drivers.opencl.scheduler;
 
+import uk.ac.manchester.tornado.api.WorkerGrid;
+import uk.ac.manchester.tornado.drivers.opencl.OCLDeviceContext;
+import uk.ac.manchester.tornado.drivers.opencl.OCLKernel;
+import uk.ac.manchester.tornado.drivers.opencl.OCLTargetDevice;
 import uk.ac.manchester.tornado.runtime.tasks.meta.TaskMetaData;
 
-public class OCLGPUScheduler extends OCLKernelScheduler {
+public class OCLAMDScheduler extends OCLKernelScheduler {
 
-    private long maxComputeUnits;
-    private long[] maxWorkGroupSize;
-
-    private static final int WARP_SIZE = 32;
+    private static final int WARP_SIZE = 64;
     private boolean ADJUST_IRREGULAR = false;
 
     private final long[] maxWorkItemSizes;
 
-    public OCLGPUScheduler(final OCLDeviceContext context) {
+    public OCLAMDScheduler(final OCLDeviceContext context) {
         super(context);
         OCLTargetDevice device = context.getDevice();
-
         maxWorkItemSizes = device.getDeviceMaxWorkItemSizes();
-        maxComputeUnits = device.getDeviceMaxComputeUnits();
-        maxWorkGroupSize = device.getDeviceMaxWorkGroupSize();
+    }
+
+    @Override
+    public int launch(long executionPlanId, OCLKernel kernel, TaskMetaData meta, int[] waitEvents, long batchThreads) {
+        if (meta.isWorkerGridAvailable()) {
+            WorkerGrid grid = meta.getWorkerGrid(meta.getId());
+            long[] global = grid.getGlobalWork();
+            long[] offset = grid.getGlobalOffset();
+            long[] local = grid.getLocalWork();
+            return deviceContext.enqueueNDRangeKernel(executionPlanId, kernel, grid.dimension(), offset, global, local, waitEvents);
+        } else {
+            return deviceContext.enqueueNDRangeKernel(executionPlanId, kernel, meta.getDims(), meta.getGlobalOffset(), meta.getGlobalWork(), null, waitEvents);
+        }
     }
 
     @Override
     public void calculateGlobalWork(final TaskMetaData meta, long batchThreads) {
         final long[] globalWork = meta.getGlobalWork();
-
         for (int i = 0; i < meta.getDims(); i++) {
             long value = (batchThreads <= 0) ? (long) (meta.getDomain().get(i).cardinality()) : batchThreads;
+            // adjust for irregular problem sizes
             if (ADJUST_IRREGULAR && (value % WARP_SIZE != 0)) {
                 value = ((value / WARP_SIZE) + 1) * WARP_SIZE;
             }
@@ -60,31 +71,27 @@ public class OCLGPUScheduler extends OCLKernelScheduler {
     @Override
     public void calculateLocalWork(final TaskMetaData meta) {
         final long[] localWork = meta.initLocalWork();
-
         switch (meta.getDims()) {
             case 3:
-                localWork[2] = 1;
-                localWork[1] = calculateGroupSize(calculateEffectiveMaxWorkItemSizes(meta)[1], meta.getGlobalWork()[1]);
-                localWork[0] = calculateGroupSize(calculateEffectiveMaxWorkItemSizes(meta)[0], meta.getGlobalWork()[0]);
-                break;
+                /// XXX: Support 3D
+                localWork[2] = calculateGroupSize(maxWorkItemSizes[2], meta.getOpenCLGpuBlock2DY(), meta.getGlobalWork()[2]);
             case 2:
-                localWork[1] = calculateGroupSize(calculateEffectiveMaxWorkItemSizes(meta)[1], meta.getGlobalWork()[1]);
-                localWork[0] = calculateGroupSize(calculateEffectiveMaxWorkItemSizes(meta)[0], meta.getGlobalWork()[0]);
+                localWork[1] = calculateGroupSize(maxWorkItemSizes[1], meta.getOpenCLGpuBlock2DY(), meta.getGlobalWork()[1]);
+                localWork[0] = calculateGroupSize(maxWorkItemSizes[0], meta.getOpenCLGpuBlock2DX(), meta.getGlobalWork()[0]);
                 break;
             case 1:
-                localWork[0] = calculateGroupSize(calculateEffectiveMaxWorkItemSizes(meta)[0], meta.getGlobalWork()[0]);
+                localWork[0] = calculateGroupSize(maxWorkItemSizes[0], meta.getOpenCLGpuBlockX(), meta.getGlobalWork()[0]);
                 break;
             default:
                 break;
         }
     }
 
-    private int calculateGroupSize(long maxBlockSize, long globalWorkSize) {
+    private int calculateGroupSize(long maxBlockSize, long customBlockSize, long globalWorkSize) {
         if (maxBlockSize == globalWorkSize) {
             maxBlockSize /= 4;
         }
-
-        int value = (int) Math.min(maxBlockSize, globalWorkSize);
+        int value = (int) Math.min(Math.max(maxBlockSize, customBlockSize), globalWorkSize);
         if (value == 0) {
             return 1;
         }
@@ -92,28 +99,5 @@ public class OCLGPUScheduler extends OCLKernelScheduler {
             value--;
         }
         return value;
-    }
-
-    private long[] calculateEffectiveMaxWorkItemSizes(TaskMetaData metaData) {
-        long[] intermediates = new long[] { 1, 1, 1 };
-
-        switch (metaData.getDims()) {
-            case 3:
-                intermediates[2] = (long) Math.sqrt(maxWorkItemSizes[2]);
-                intermediates[1] = (long) Math.sqrt(maxWorkItemSizes[1]);
-                intermediates[0] = (long) Math.sqrt(maxWorkItemSizes[0]);
-                break;
-            case 2:
-                intermediates[1] = (long) Math.sqrt(maxWorkItemSizes[1]);
-                intermediates[0] = (long) Math.sqrt(maxWorkItemSizes[0]);
-                break;
-            case 1:
-                intermediates[0] = maxWorkItemSizes[0];
-                break;
-            default:
-                break;
-
-        }
-        return intermediates;
     }
 }
