+++ tornado-drivers/opencl/src/main/java/uk/ac/manchester/tornado/drivers/opencl/scheduler/OCLFPGAScheduler.java
@@ -23,12 +23,14 @@
  * Authors: Michalis Papadimitriou
  *
  */
-package uk.ac.manchester.tornado.drivers.opencl;
-
-import uk.ac.manchester.tornado.api.WorkerGrid;
-import uk.ac.manchester.tornado.runtime.tasks.meta.TaskMetaData;
+package uk.ac.manchester.tornado.drivers.opencl.scheduler;
 
 import java.util.Arrays;
+
+import uk.ac.manchester.tornado.api.WorkerGrid;
+import uk.ac.manchester.tornado.drivers.opencl.OCLDeviceContext;
+import uk.ac.manchester.tornado.drivers.opencl.OCLKernel;
+import uk.ac.manchester.tornado.runtime.tasks.meta.TaskMetaData;
 
 public class OCLFPGAScheduler extends OCLKernelScheduler {
